+++ libs/common/src/lib/entities/add-ner-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './add-ner-dialog.component';